Fix aspnet classic tests
